Move Error/Result to own mod
